Free-Code-Camp-Projects-gemini/ATS/requirements.txt
streamlit
google-generativeai
python-dotenv
PDF2Image